# [prototype] Improve performance of liftstd