@fingolfin 2021-01-17
This sounds pretty cool. I wonder if @hannes14 has any thoughts on this?