@hannes14 2021-01-18
Impressive and very interessting - I appriciate the ideas behind this patch,
but have problems with the real patch:
- it nearly doubles the size of TObject (nad LObject) in best case, increase the size in the worst case by 16 kB (and we may have several 100 000 of them)
- test fail for every use of Groebner bases except pure std and liftstd
(for example intersection, quotient, resolution).

I am working on integration of the good ideas, which I see a (in the case of liftstd):
- stop to work on polynomials which turn out to be a syzygy as soon as possible
  (do not enter ist in the list of reducesrs, nad partial GB)
  This is already succesful (08dc13b40d19af0e4a91ab3cd7d095b5884920b7) and 
   reduces compution time by 50 % (in the example)
-  do the reduction as "lazy computation" (as lazy as possible) - most of the reductions tend to 
   lead to syzygies (which are not needed for liftstd). This is more difficult - 
   not just "simple book keeping".