@zickgraf 2021-01-18
> but have problems with the real patch:
> 
>     * it nearly doubles the size of TObject (nad LObject) in best case, increase the size in the worst case by 16 kB (and we may have several 100 000 of them)
Yes, of course that's not the way the final solution should look like. I tried to attach the data to `strat` instead (see ffb22cef4dda29c96628d79fdda9b0232e73017c), which I assumed would be the canonical place to store the data, but my C skills were not sufficient to get the memory management correct.
> 
>     * test fail for every use of Groebner bases except pure std and liftstd
>       (for example intersection, quotient, resolution).
Yes, that's expected due to all the edge cases that are missing. In fact, I am surprised that the `liftstd` tests succeed :D
> 
> 
> I am working on integration of the good ideas

Thanks! I will report if I see improvements in my computations :-)

> , which I see a (in the case of liftstd):
> 
>     * stop to work on polynomials which turn out to be a syzygy as soon as possible
>       (do not enter ist in the list of reducesrs, nad partial GB)
>       This is already succesful ([08dc13b](https://github.com/Singular/Singular/commit/08dc13b40d19af0e4a91ab3cd7d095b5884920b7)) and
>       reduces compution time by 50 % (in the example)
I have used this optimization during my tests (its the suggestion I made here: https://github.com/Singular/Singular/pull/965#issuecomment-562643255), but I think this only disguises the problem. If I understand this correctly, one could even remove `IDLIFT` when computing the transformation matrix and syzygies lazily, as then they would be kept in a separate data structure anyway.
> 
>     * do the reduction as "lazy computation" (as lazy as possible) - most of the reductions tend to
>       lead to syzygies (which are not needed for liftstd). This is more difficult -
>       not just "simple book keeping".
Maybe we have different things in mind when talking about "book keeping". For me, "book keeping" by definition includes everything I do to make computations lazy. I will avoid the term in the future.