@zickgraf 2021-04-19
@hannes14 Thank you very much for turning this prototype into production code!!! With 5f5ce616bcb84ae7ec4aa40b04675fe7a3e7e84c, `liftstd` seems to be indeed much faster 🎉 Unfortunately, with 817917a7da7faac058cd133c5aef74092b4cf6af we are back at the original speed, so I assume something in that commit is breaking the laziness again :/