@hannes14 2021-04-19
should be fixed with https://github.com/Singular/Singular/commit/ddebd75a4ce9e60424c1454a22dee815afe211fe