@zickgraf 2021-04-20
That commits does the trick, thanks a lot for the quick fix!

Now I have tried to use it with the relative version of `liftstd` and get a segfault with the following example:
```
LIB "matrix.lib";

ring R = 0,(x,y,z),(c,dp);

matrix A[1][1] = x;
matrix B[1][2] = x+y-z,x-y-z;

matrix C[nrows(A)][ncols(A) + ncols(B)] = concat(A,B);

matrix G1 = std(C);

matrix T;
matrix G2 = liftstd(module(A), T, "std", module(B));

print("");
print("reduce(G1,G2):");
print(reduce(G1,G2));
print("reduce(A*T-G2,std(B)):");
print(reduce(A*T-G2,std(B)));
```
I could fix the segfault (https://github.com/zickgraf/Singular/commit/b114da643beb4812d859b81b7fd9e8d6fa41e9e6) but with the fix I get assume violations:
```
// ***dError: assume violation at ../../../../kernel/GBEngine/kspoly.cc:309 condition: PW->GetpLength() == pLength(PW->p != __null ? PW->p : PW->t_p)
// ***dError: assume violation at ../../../../kernel/GBEngine/kspoly.cc:309 condition: PW->GetpLength() == pLength(PW->p != __null ? PW->p : PW->t_p)
// ***dError: assume violation at ../../../../kernel/GBEngine/kspoly.cc:309 condition: PW->GetpLength() == pLength(PW->p != __null ? PW->p : PW->t_p)
```
and wrong results (the two reductions at the end should both yield 0).