@hannes14 2021-04-20
Thanks for the example.
Fixed with 30e184b9f651e8205529c6336cd2f7c39e603699.
Assume violations: in ksReducePoly, only the first part of the polynomials occur,
therefore the length was wrong.