@zickgraf 2021-04-22
Again, many thanks for the quick fix! Now my applications look very promising :-)

Unfortunately, I'm running into a new segfault due to `p` being `NULL` here: https://github.com/Singular/Singular/blob/5fd24e69d9cdfcd1040ca93af529563f039efd2c/kernel/GBEngine/kLiftstd.cc#L83

Minimal example:
```
LIB "matrix.lib";

ring R = 0,(x,y),(c,dp);

matrix A[2][1] = x, x;
matrix B[2][1] = y, y;

matrix C[nrows(A)][ncols(A) + ncols(B)] = concat(A,B);

matrix G1 = std(C);

matrix T;
matrix G2 = liftstd(module(A), T, "std", module(B));
```

And I have noticed something weird about the runtime: when choosing large random matrices `A` and `B` in the example above, `liftstd` is actually a bit faster than `std`, which is quite unexpected for me:
```
LIB "random.lib";
// make random reproducible
system("random", 1);
system("--ticks-per-sec",1000);

int t;

int rows = 30;
int columns1 = 350;
int columns2 = 350;

ring R = 0,(x,y,z),(c,dp);
matrix A = randommat(rows, columns1, maxideal(1), 10);
matrix B = randommat(rows, columns2, maxideal(1), 10);
A[1,1] = x^2;

matrix C[nrows(A)][ncols(A) + ncols(B)] = concat(A,B);

t = timer;
matrix G1 = std(C);
print("computed std(C) in " + string(timer - t) + " ms");

matrix T;
t = timer;
matrix G2 = liftstd(module(A), T, "std", module(B));
print("computed liftstd in " + string(timer - t) + " ms");
```
produces
```
computed std(C) in 2220 ms
computed liftstd in 2090 ms
```

But at least until now I could not find an example producing wrong results 🤔 Maybe this is also connected to the segfault.