@zickgraf 2021-04-27
> segfault fixed with 10c7e04

Thanks! And sorry for always bothering you. Usually I try to fix such issues myself, but with these `NULL`-pointer accesses I am never sure if I should simply return `NULL` or if they need some additional handling.